+++ ideas/package.json
@@ -11,7 +11,7 @@
   "jest": {
     "preset": "ts-jest",
     "testEnvironment": "node",
-    "setupFilesAfterEnv": ["./src/test/setup.ts"]
+    "setupFilesAfterEnv": ["./src/__test__/setup.ts"]
   },
   "dependencies": {
     "@ije-ticketapp/common": "^1.0.18",
