+++ ideas/src/idea/idea.spec.ts
@@ -1,5 +1,5 @@
 import { ScoreTypes } from './idea'
-import { generateMakeIdeaProps } from './../test/test-utils'
+import { generateMakeIdeaProps } from '../__test__/test-utils'
 import { makeIdea, Id } from '.'
 
 describe('idea', () => {
